Refactor and improve error handling in SpiceClient and SpiceClientShared
- Changed `unwrap_or` to `first` for better readability when extracting host from URI.
- Simplified error logging by using the `{}` syntax for formatting in multiple locations.
- Updated documentation comments to reflect the need for future documentation.
- Adjusted the handling of mouse wheel events to ignore unused parameters.
- Enhanced the handling of connection errors and task join errors for clarity.
- Removed unnecessary imports and cleaned up code in multimedia modules.
- Improved test cases for cursor and display channels to use `bincode` serialization.
- Added default implementation for `TestHarness` struct to simplify test setup.
- Ensured consistent error handling across transport and video output traits.

+++ gtk4-app/src/ui/vm_create_dialog.rs
@@ -390,7 +390,7 @@ impl VMCreateDialog {
 
         let version = if let Some(model) = imp.version_row.model() {
             if let Some(string_list) = model.downcast_ref::<gtk::StringList>() {
-                let version_index = imp.version_row.selected() as u32;
+                let version_index = imp.version_row.selected();
                 if version_index < string_list.n_items() {
                     string_list.string(version_index).map(|s| s.to_string())
                 } else {
@@ -406,7 +406,7 @@ impl VMCreateDialog {
         let edition = if imp.edition_row.is_visible() {
             if let Some(model) = imp.edition_row.model() {
                 if let Some(string_list) = model.downcast_ref::<gtk::StringList>() {
-                    let edition_index = imp.edition_row.selected() as u32;
+                    let edition_index = imp.edition_row.selected();
                     if edition_index < string_list.n_items() {
                         string_list.string(edition_index).map(|s| s.to_string())
                     } else {
